backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Bcrypt==1.0.1
pymongo==4.5.0
dnspython==2.4.2
bcrypt==4.0.1
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
qiskit==0.44.1
qiskit-aer==0.12.2
reportlab==4.0.4
Pillow==10.0.0
pytest==7.4.2
pytest-cov==4.1.0
gunicorn==21.2.0
psutil==5.9.5
SQLAlchemy==2.0.21
Werkzeug==2.3.7
bcrypt==4.0.1
marshmallow==3.20.1
